backend/requirements/base.txt
dj-database-url==2.3.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
Markdown==3.7
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
